int.c size_t i

+++ aula03/int.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 #include <limits.h>
 #include <stdint.h>
-
+#include <stdlib.h>
 int main(void){
-  long long i;
+  size_t i;
 
   printf("O tamanho de i (int): %zu bytes %zu bits ",sizeof i, sizeof i * 8);
-  printf("\nValor de i:%lld\n",i);
+  printf("\nValor de i:%zu\n",i);
 
   return 0;
 }
